% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_frequency_plot.R
\name{aa_frequency_plot}
\alias{aa_frequency_plot}
\title{Amino Acids Frequency Plot}
\usage{
aa_frequency_plot(aa_chain)
}
\arguments{
\item{aa_chain}{a chain of amino acids as a string}
}
\value{
Show a plot which count how many times each AA appear in the amino acids chain given as input
}
\description{
Amino Acids Frequency Plot
}
\examples{
if (interactive()){
      aa_chain <- "AWLPPHWLLQRI"
      aa_frequency_plot(aa_chain)
      }
}
