% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_distribution.R, R/ammino_distribution.R
\name{amino_distribution}
\alias{amino_distribution}
\title{Amino acids distribution}
\usage{
amino_distribution(amino_sequence)

amino_distribution(amino_sequence)
}
\arguments{
\item{amino_sequence}{String containing an sequence of amino acids}
}
\value{
Bar diagram showing the distribution of amino acids in the input string

Bar diagram showing the distribution of amino acids in the input string
}
\description{
Amino acids distribution

Amino acids distribution
}
\examples{
amino_sequence <- "AGCTAGCTAGCTAGC"
amino_plot <- amino_distribution(amino_sequence)
amino_sequence <- "AGCTAGCTAGCTAGC"
amino_plot <- amino_distribution(amino_sequence)
}
