% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammino_distribution.R
\name{ammino_distribution}
\alias{ammino_distribution}
\title{Amminoacids distribution}
\usage{
ammino_distribution(ammino_sequence)
}
\arguments{
\item{ammino_sequence}{String containing an sequence of amminoacids}
}
\value{
Bar diagram showing the distribution of amminoacids in the input string
}
\description{
Amminoacids distribution
}
\examples{
ammino_sequence <- "AGCTAGCTAGCTAGC"
ammino_plot <- ammino_distribution(ammino_sequence)
}
