% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna.R
\name{build_dna}
\alias{build_dna}
\title{Title
Makes a DNA sequence from a string of nucleotides}
\usage{
build_dna(sequence_length)
}
\arguments{
\item{sequence_lenght}{an integer to indicate the number of nucleotides used, final lenght of the dna sequence}
}
\value{
a string of nucleotides
}
\description{
Title
Makes a DNA sequence from a string of nucleotides
}
\examples{

sequence_lenght <- 12
build_dna(sequence_lenght)
}
