% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_aa.R
\name{codon_to_aa}
\alias{codon_to_aa}
\title{Translation of codons into amino acids}
\usage{
codon_to_aa(codons)
}
\arguments{
\item{codons}{List of codons to translate into their amino acid sequence}
}
\value{
Amino acid sequence corresponding to the codon list
}
\description{
`codon_to_aa()` translates a codon sequence into an amino acid sequence
}
