% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_transcription.R
\name{dna_transcription}
\alias{dna_transcription}
\title{DNA transcription}
\usage{
dna_transcription(dna_sequence)
}
\arguments{
\item{dna_sequence}{sequence of DNA to be transcripted}
}
\value{
sequence of corresponding RNA
}
\description{
DNA transcription
}
\examples{
dna_sequence <- c("TTT", "TCT", "TAT")
dna_sequence <- toupper(dna_sequence)
transcripted_rna <- gsub("T", "U", dna_sequence)
}
