% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Title: Split an RNA sequence into codons}
\usage{
extract_codons(rna_sequence, start = 1)
}
\arguments{
\item{start}{indicates the starting position withing the RNA sequence (default is 1)}

\item{rna_sequence:}{string of RNA nucleotides}
}
\value{
codons: groups of three nucleotides each; each codon correspons to an amino acid or stop signal
}
\description{
Title: Split an RNA sequence into codons
}
\examples{
extract_codons("AUGGCUUAA")

extract_codons("AUGGCUUAA", start = 2)

}
