% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extracted_codons}
\alias{extracted_codons}
\title{extract_codons}
\usage{
extracted_codons(RNA_string, start = 1)
}
\arguments{
\item{RNA_string}{RNA string from which to extract the codons.}

\item{start}{Starting position to start extracting the first codon. It defines the ORF (Open Reading Frame)}
}
\value{
Extracted codons, given the RNA sequence and the ORF chosen.
}
\description{
extract_codons
}
