% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DNA_seq.R
\name{generate_DNA_seq}
\alias{generate_DNA_seq}
\title{DNA sequence generator.}
\usage{
generate_DNA_seq(seq_len)
}
\arguments{
\item{seq_len}{Length of the DNA sequence that will be generated}
}
\value{
The function returns the sequence of nucleotides that conform the DNA
}
\description{
`generate_DNA_seq()` generates a random DNA sequence of the desired length.
}
