% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DNA_sequence.R
\name{generate_dna_sequence}
\alias{generate_dna_sequence}
\title{Generate DNA sequence}
\usage{
generate_dna_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{length of the DNA sequence you want to create}
}
\value{
returns the random DNA sequence of length 'sequence_length'
}
\description{
Generate DNA sequence
}
\examples{
sequence_length <- 10
DNA_sequence <- generate_dna_sequence(sequence_length)
}
