% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_protein_sequence.R
\name{generate_protein_sequence}
\alias{generate_protein_sequence}
\title{Generate protein sequence}
\usage{
generate_protein_sequence(vect_codons)
}
\arguments{
\item{vect_codons}{vector of codons (e.g. c("AUG", "UUU", "CGC"))}
}
\value{
returns the sequence of AA that correspond to the input codons
}
\description{
Generate protein sequence
}
\examples{
vect_codons <- c("AUG", "UUU", "CGC")
protein_sequence <- generate_protein_sequence(vect_codons)
}
