% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_codons.R
\name{make_codons}
\alias{make_codons}
\title{Convert a RNA sequence into codons}
\usage{
make_codons(rna_seq, start = 1)
}
\arguments{
\item{rna_seq}{The input RNA sequence that is going to be converted to codons}

\item{start}{First position to start making the codons}
}
\value{
A vector of codons derived from the RNA sequence
}
\description{
Convert a RNA sequence into codons
}
\examples{
make_codons("AUGCGAUGC", start = 1)
#> "AUG" "CGA" "UGC"
}
