% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_counts.R
\name{plot_aa_counts}
\alias{plot_aa_counts}
\title{Plot amino acid counts}
\usage{
plot_aa_counts(seq)
}
\arguments{
\item{seq}{Sequence of amino acids whose amino acid frequency we want to determine}
}
\value{
The function returns a plot with the amino acid frequency of the input sequence
}
\description{
`plot_aa_counts()` determines the frequency of each amino acid in the input sequence and represents it in a bar plot.
}
