% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcription of DNA into RNA}
\usage{
transcribe(DNA_seq)
}
\arguments{
\item{DNA_seq}{DNA sequence to be transcribed}
}
\value{
RNA sequence transcribed from DNA sequence
}
\description{
'transcribe()' transcribes a given DNA sequence into an RNA sequence
}
