% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription}
\usage{
transcription(dna_seq)
}
\arguments{
\item{dna_seq}{character sequence to transcribe into RNA}
}
\value{
RNA sequence where all Thymine have been converted into Uracil
}
\description{
Transcription
}
\examples{
transcription('ACACGTAGTTA')
}
