% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation
Translate mRNA codons into the corresponding amino acids chain}
\usage{
translation(codons)
}
\arguments{
\item{codons}{a vector containing codons as strings}
}
\value{
a single string containing all the amino acids
}
\description{
Translation
Translate mRNA codons into the corresponding amino acids chain
}
\examples{

codons <- c("AUG", "CCU","CAA", "CGA", "UGG", "GGC")
translation(codons)

}
